requirements.txt
streamlit==1.30.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
pytest==7.4.3
fpdf2==2.7.8
reportlab==4.2.0
Pillow==10.3.0